Dev #348: initial refactor to lift out monolithically organised methods in icenet.plotting.forecast into specific modules (CLI, process.compute, etc...) where easily achieved
