+++ pyproject.toml
@@ -64,14 +64,16 @@ icenet_dataset_check = "icenet.data.network_dataset:check_dataset"
 icenet_dataset_create = "icenet.data.loader:create_network_dataset"
 icenet_generate_ref_osisaf = "icenet.data.references.osisaf:generate"
 icenet_plot_record = "icenet.plotting.data:plot_tfrecord"
-icenet_plot_forecast = "icenet.plotting.forecast:plot_forecast"
 icenet_plot_input = "icenet.plotting.data:plot_sample_cli"
-icenet_plot_sic_error = "icenet.plotting.forecast:sic_error"
-icenet_plot_sic_error_local = "icenet.plotting.forecast:sic_error_local"
-icenet_plot_bin_accuracy = "icenet.plotting.forecast:binary_accuracy"
-icenet_plot_sie_error = "icenet.plotting.forecast:sie_error"
-icenet_plot_metrics = "icenet.plotting.forecast:metric_plots"
-icenet_plot_leadtime_avg = "icenet.plotting.forecast:leadtime_avg_plots"
+
+icenet_plot_forecast = "icenet.plotting.forecast:forecast_cli"
+icenet_plot_sic_error = "icenet.plotting.forecast:sic_error_cli"
+icenet_plot_sic_error_local = "icenet.plotting.forecast:sic_error_local_cli"
+icenet_plot_bin_accuracy = "icenet.plotting.forecast:binary_accuracy_cli"
+icenet_plot_sie_error = "icenet.plotting.forecast:sie_error_cli"
+icenet_plot_metrics = "icenet.plotting.forecast:metric_cli"
+icenet_plot_leadtime_avg = "icenet.plotting.forecast:leadtime_avg_cli"
+
 icenet_predict = "icenet.model.predict:main"
 icenet_output = "icenet.process.predict:create_cf_output"
 icenet_output_geotiff = "icenet.process.forecasts:create_geotiff_output"
